Move methods and trait over to AbstractProductType - refactor later
